Initial commit
pycsdl2 is a low-overhead, compiled SDL2 binding for CPython. It aims to
provide the familiar C API of SDL2, while offering several Pythonic
features such as automatic memory management, bounds checking and
exceptions.

The source code is licensed under the zlib license.

We begin with a simple README, LICENSE, and gitignore. Hello world!