setup.py: introduce get_csdl2_base_ext()
Since the csdl2 extension will need to link to SDL2, separate the
extension source files definition from any future link flags definition
code by introducing get_csdl2_base_ext(), which returns the csdl2
Extension that is not linked to SDL2. Thus, to link the extension to
SDL2, we simply need to modify the properties of the returned Extension
object in a separate function.