setup.py: implement get_csdl2_bundled_ext()
In operating systems like Windows and Mac OS X where the environment is
generally well controlled and stable, it is useful to be able to build
and static link pycsdl2 to the bundled SDL2 source code so as to
minimize errors due to missing/wrong version of libraries/header files
and to make it easier to build and distribute binary packages of
pycsdl2.

As such, implement building and linking the bundled SDL2 source code on
Windows and Mac OS X by introducing get_csdl2_bundled_ext(). It is only
possible to build the bundled source code for these operating systems
because SDL ships with config headers for them in SDL_config_windows.h
and SDL_config_macosx.h respectively.

The list of source globs and libraries to link against were taken from
SDL2's CMakeLists.txt and tested on Windows 7 and Mac OS X Yosemite.

In addition, since get_csdl2_bundled_ext() does not support other
operating systems, we first try get_csdl2_bundled_ext(), and if it
fails with NotImplementedError, we fall back on get_csdl2_system_ext().