test: add build/lib.* to sys.path
distutils outputs the built csdl2 extension in
'build/lib.{platform}-{version}', where `platform` is the value of
distutils.util.get_platform(), and `version` is the value of
sys.version[0:3] (See the source code of distutils.command.build). This
built extension is what we want to import and test, so add this path to
sys.path before importing or running the tests.