distutils.h: implement PyCSDL2_GetSystemSDL()
Sometimes, extensions want to link against the same system SDL2 library
which pycsdl2 has linked against. It would be useful to expose the
compile and link flags that were used to compile the csdl2 extension so
that other extensions can use the same flags to link against the same
library. Thus, implement PyCSDL2_GetSystemSDL() which will provide the
above information in the format which is friendly to construct a
distutils.extension.Extension from.