test_distutils: verify return value of PyCSDL2_GetSystemSDL()
Check to see if the flags returned by PyCSDL2_GetSystemSDL() are
actually valid information.

Since the main purpose of implementing PyCSDL2_GetSystemSDL() is to
allow other extensions to link against the same system SDL2 library as
the csdl2 extension, it would be useful to test this use case as well.

Implement a test that builds an extension that links against system SDL2
using the compile and link flags returned by PyCSDL2_GetSystemSDL(), and
then run it by importing it. SDL_Log() is chosen because it can
successfully run on headless systems.