render.h: implement PyCSDL2_RendererValid()
Asserts on SDL_Renderer only checked if the internal SDL_Renderer is not
NULL, but did not check whether the SDL_Renderer's underlying device
(e.g. SDL_Window or SDL_Surface) is not NULL.

Fix this by introducing a PyCSDL2_RendererValid() function which will do
the correct checks.

Furthermore, it should throw ValueError on invalid SDL_Renderer's, not
AssertionError. This is keeping in line with what memoryview throws when
it has been released.