rwops: remove PyCSDL2_RWopsPtr, simplify code
PyCSDL2_RWopsPtr was introduced in 68022a4 (rwops.h: introduce
PyCSDL2_RWopsPtrType, 2015-04-22) because it was thought that csdl2's
SDL_RWops interface may need to provide a "view" into its "hidden"
union, and so we need an internal object to "own" the SDL_RWops pointer
and share it between the different "views".

However, it turns out we are not going to need this, as it is dangerous
to let the user modify the "hidden" union. As such, we don't need the
PyCSDL2_RWopsPtr type, so we can remove it, and simplify the code as a
result.