util.h: introduce SIZE_T_UNIT
CPython has the format unit "n" which corresponds to the Py_ssize_t data
type. However, when we want a size_t argument, it is better to use the
unsigned variant, as we get overflow checking (e.g. the argument should
not be negative).

As such, we define the macro SIZE_T_UNIT, which corresponds to the C
unsigned integer format unit with the same width as size_t.