Test_SDL_RWFromFile: replace with TestRWFromFile_{Read,Write}
Almost every test in Test_SDL_RWFromFile has a:

    rw = SDL_RWFromFile(file, mode)

at the beginning of the test, which is a lot of repetition. The problem
is that we mixed tests for SDL_RWFromFile(file, 'r') and
SDL_RWFromFile(file, 'w') in the same TestCase, so we can't take
advantage of setUp()

As such, split up the SDL_RWFromFile(file, 'r') tests and
SDL_RWFromFile(file, 'w') tests into their own test cases.