PyCSDL2_RWops: implement PyCSDL2_RWopsFinalize
In PyCSDL2_RWopsDealloc(), we attempt to call the close() callback of
the SDL_RWops so as to prevent file descripter leakage.

However, the close() callback may Py_INCREF() the PyCSDL2_RWops object,
thus ressurecting it, or it may Py_INCREF() and then Py_DECREF() the
PyCSDL2_RWops object, hence leading to an infinite
PyCSDL2_RWopsDealloc() recursion.

As such, we take advantage of the tp_finalize slot, introduced in Python
3.4. It is called before the garbage collector, so the object is
guaranteed to be in a sane state, and also detects if the object was
resurrected.