rwops: implement PyCSDL2_RWsize(), replacing PyCSDL2_RWSizeFuncCall()
The function signatures of PyCSDL2_RWsize() and
PyCSDL2_RWSizeFunc.__call__() are exactly the same -- the only
difference is that the former gets the callback pointer from the
provided SDL_RWops, while the latter gets the callback pointer from its
own instance data.

As such, we rename PyCSDL2_RWSizeFuncCall() to PyCSDL2_RWsize(), and
then tweak the function body so we support the above two APIs.