rwops: implement PyCSDL2_RWread(), replacing PyCSDL2_RWReadFuncCall()
The function signatures of PyCSDL2_RWread() and
PyCSDL2_RWReadFunc.__call__() are exactly the same -- the only
difference is that the former gets the callback pointer from the
provided SDL_RWops, while the latter gets the callback pointer from its
own instance data.

As such, we rename PyCSDL2_RWReadFuncCall() to PyCSDL2_RWread(), and
then tweak the function body so we support the above two APIs.