rwops: implement PyCSDL2_RWwrite(), replacing PyCSDL2_RWWriteFuncCall()
The function signatures of PyCSDL2_RWseek() and
PyCSDL2_RWSeekFunc.__call__() are exactly the same -- the only
difference is that the former gets the callback pointer from the
provided SDL_RWops, while the latter gets the callback pointer from its
own instance data.

As such, we rename PyCSDL2_RWWriteFuncCall() to PyCSDL2_RWwrite(), and
then tweak the function body so we support the above two APIs.