rwops: implement PyCSDL2_RWclose(), replacing PyCSDL2_RWCloseFuncCall()
The function signatures of PyCSDL2_RWclose() and
PyCSDL2_RWCloseFunc.__call__() are exactly the same -- the only
difference is that the former gets the callback pointer from the
provided SDL_RWops, while the latter gets the callback pointer from its
own instance data.

As such, we rename PyCSDL2_RWCloseFuncCall() to PyCSDL2_RWclose(), and
then tweak the function body so we support the above two APIs.