PyCSDL2_SurfacePixels: use PyCSDL2_SurfacePixelsValid()
Currently, we use

    PyCSDL2_Assert(self->surface, -1);

to check to see if there is a valid SDL_Surface attached to the
PyCSDL2_SurfacePixels object.

However, in addition, we need to assert that the "pixels" field is
initialized to an approriate buffer as well.

As such, refactor the PyCSDL2_Assert() into a single
PyCSDL2_SurfacePixelsValid() function to validate the
PyCSDL2_SurfacePixels object.