PyCSDL2_SurfaceRect: use PyCSDL2_SurfaceRectValid()
All over src/surface.h, we use

    PyCSDL2_Assert(self->surface, NULL);

to check to see if the PyCSDL2_SurfaceRect object has a valid
SDL_Surface attached to it.

However, this does not scale well. Should we introduce a new field, we
would have to add lots of additional PyCSDL2_Assert()s to check for it.

As such, refactor all the checks to a single function,
PyCSDL2_SurfaceRectValid().

Furthermore, check that the "rect" field is pointing to a valid SDL_Rect
as well.