util.h: introduce PyCSDL2_Buffer and its protocol functions
In some parts of the SDL API, csdl2 has to expose a block of memory
which is managed by SDL. This means that we cannot use our memory
management routines, as we, or the SDL library, may be using a different
memory manager.

While we could use PyMemoryView_FromMemory() to expose this memory,
there are some problems:

1. We cannot attach a destructor to the PyMemoryView, which means we
   will leak the memory.

2. PyMemoryView does not tell us if it exported the memory pointer to
   other objects. This means we do not know if it is safe to tell SDL to
   free that piece of memory or not, as there may be other python users
   pointing to that block of memory.

As such, we have to create our own types to expose that block of memory.
However, this would also mean that we have to re-implement
__getitem__(), __setitem__(), __len__() etc.

To make it easier, we introduce the PyCSDL2_Buffer*() utility protocol
functions that help us implement these protocol functions, provided that
our instance data contains PyCSDL2_BufferHEAD.