rwops: introduce PyCSDL2_RWopsDetach()
PyCSDL2_RWops objects are invalidated by doing:

    rwops_obj->rwops = NULL;

However, in addition to that, we need to clear all our references to
Python objects as well, as we do not need them anymore.

As such, introduce the PyCSDL2_RWopsDetach() function, that not only
sets the "rwops" field to NULL, but calls PyCSDL2_RWopsClear() as well.