set version number to 2.0.0.0.dev0
Switch pycsdl2 to a new version numbering scheme:

    SDL_VERSION.PYCSDL2_VERSION[.devX]

Where SDL_VERSION is the SDL library version that the bindings are based
on and expect (currently 2.0.0), and PYCSDL2_VERSION is a single
component version number that is incremented every time a new release is
made.

I also plan to release regular development snapshots, and they will be
identified by the .devN suffix, where N is a number that is incremented
with every new development release.