PyCSDL2_BufferCreate(): optionally keep a reference to the owner
While PyCSDL2_Buffer is highly convenient for exposing arrays that are
part of a struct, it depends on the Python object wrapping the struct
memory to exist, to ensure that the array's memory does not go away.

We can model this dependency by allowing PyCSDL2_Buffer to optionally
keep a reference to a PyObject that manages the buffer, to ensure that
the buffer exists for the PyCSDL2_Buffer's lifetime.