util: introduce PyCSDL2_PyModuleAddType()
We often do the following to add a type object to the module:

    if (PyType_Ready(&type)) { return 0; }
    Py_INCREF(&type);
    if (PyModule_AddObject(module, "The_Type_Name", (PyObject *)&type))
        return 0;

This contains a lot of boilerplate -- we need to repeat the type name
multiple times.

Instead, move the boilerplate to the utility function
PyCSDL2_PyModuleAddType(), so we just need to write the following:

    if (PyCSDL2_PyModuleAddType(module, &type) < 0)
        return 0;

which is a lot shorter and succinct.