PyCSDL2_Event: make exposed buffer opaque
It does not make sense to allow byte-level access to the buffer of the
SDL_Event, as they are not bytes. Instead, we make the exposed buffer
format richer, exposing the "type" and "timestamp" fields, and leaving
the rest as padding bytes.

This is also so that SDL_Event can be used for the PyCSDL2_ARRAYVIEW
functionality when we introduce it, as it requires that ndim == 1, and
itemsize == sizeof(SDL_Event).